auth jwt bcrypt tests

+++ controllers/todo-controllers/deleteTodo.js
@@ -1,8 +1,9 @@
 import Todo from '../../models/todo-model.js';
+import User from '../../models/user-model.js';
 import { redisClient } from '../../db/redisClient.js';
 
 const deleteTodo = async (req, res) => {
-  // const userId = req.userId;
+  const userId = req.userId;
   const todoId = req.params.todoId;
   console.log('Received todoId:', todoId);
 
@@ -15,12 +16,29 @@ const deleteTodo = async (req, res) => {
     throw error;
   }
 
-  // Redis
-  await redisClient.del(`todo:${todoId}`);
-  console.log('Deleted todo from Redis cache');
+  const user = await User.findById(userId);
+
+  if (!user) {
+    const error = new Error('User not found!');
+    error.statusCode = 404;
+    throw error;
+  }
+
+  if (user.role !== 'admin' && deletedTodo.userId.toString() !== userId) {
+    const error = new Error('Not authorized!');
+    error.statusCode = 403;
+    throw error;
+  }
 
   await Todo.deleteOne({ _id: todoId });
   console.log('Deleted todo from MongoDB');
+  // await user.updateOne({ $pull: { todos: deletedTodo._id } });
+  await user.todos.pull(deletedTodo._id);
+  await user.save();
+
+  // Redis
+  await redisClient.del(`todo:${todoId}`);
+  console.log('Deleted todo from Redis cache');
 
   res
     .status(200)
